basic server using flask
This is a remote server, used to watch keystrokes logged from any target device. Make sure the server is configured to `http://0.0.0.0:5000/` so that it is configured to device IPv4 address locally. `/` is the landing page which tells that our server is running, `/capture` route captures all the exfiltrated data from exfil.py. `/view` route shows all the decrypted data(keystrokes) from the target device.